Reorder function pointer tests to support rest parameter tests
