+++ tests/jballerina-unit-test/src/test/resources/test-src/expressions/lambda/function-pointers-with-optional-args-negative.bal
