+++ tests/jballerina-unit-test/src/test/resources/test-src/expressions/lambda/function-pointers-with-rest-args-negative.bal
@@ -1,8 +1,7 @@
-function funcWithRestArgs (int a, int b, int... c) returns [int, int, int[]] {
-    return [a, b, c];
+public function testFunctionPointerAssignmentWithRestParamsNegative() {
+   function (string a, int[] b) returns string bar = foo;
 }
 
-function test1() returns [int, int, int[]]{
-    function (int, int, int...) returns [int, int, int[]] func = funcWithRestArgs;
-    return func(1, 2, 3, 4);
+function foo(string c, int... d) returns string {
+    return c;
 }
