+++ tests/jballerina-unit-test/src/test/resources/test-src/expressions/lambda/function-pointers.bal
@@ -160,20 +160,6 @@ function testSubTypingWithAny() returns any {
     return s;
 }
 
-public function testFunctionPointerRest() returns int[]{
-   var bar = foo;
-   return bar("hello", 1, 2, 3);
-}
-
-public function testFunctionPointerRestTyped() returns int[]{
-   function (string b, int... c) returns int[] bar = foo;
-   return bar("hello", 4, 5, 6);
-}
-
-function foo(string b, int... c) returns int[] {
-    return c;
-}
-
 var f1 = function() returns string {
     return "f1";
 };
